feat(migration): add tools
